Move hit selector code into a separate module.
